Fix calculation of __length_hint__ on chord.tasks when it's a group
Before group.__iter__() was removed in 8c7ac5d, passing either a list or
a group through list() would yield the same thing: the list of tasks.
After the removal of group.__iter__(), list(g) yields the same thing as
g.keys(), so we need to handle the group case explicitly now.

Using len(g.keys()) for __length_hint__ will only happen to be the
correct chord size sometimes; the rest of the time it will result in
chords that never get unlocked due to a mismatch between the computed
chord size and the size of the complete set of results returned.

Fixes celery/celery#3725